+++ src/pharmpy/modeling/plots.py
@@ -5,15 +5,19 @@ from pharmpy.deps import pandas as pd
 from .data import get_observations
 
 
-def plot_iofv_vs_iofv(model, other):
+def plot_iofv_vs_iofv(iofv1, iofv2, name1, name2):
     """Plot individual OFV of two models against each other
 
     Parameters
     ----------
-    model : Model
-        The first model
-    other : Model
-        The second model
+    iofv1 : pd.Series
+        Estimated iOFV of the first model
+    iofv2 : pd.Series
+        Estimated iOFV of the second model
+    name1 : str
+        Name of first model
+    name2 : str
+        Name of second model
 
     Results
     -------
@@ -21,12 +25,12 @@ def plot_iofv_vs_iofv(model, other):
         Scatterplot
 
     """
-    x_label = f'{model.name} iOFV'
-    y_label = f'{other.name} iOFV'
+    x_label = f'{name1} iOFV'
+    y_label = f'{name2} iOFV'
     df = pd.DataFrame(
         {
-            x_label: model.modelfit_results.individual_ofv,
-            y_label: other.modelfit_results.individual_ofv,
+            x_label: iofv1,
+            y_label: iofv2,
         }
     )
     id_name = df.index.name