+++ tests/modeling/test_plots.py
@@ -3,7 +3,8 @@ from pharmpy.modeling import plot_individual_predictions, plot_iofv_vs_iofv
 
 def test_plot_iofv_vs_iofv(load_example_model_for_test):
     model = load_example_model_for_test('pheno')
-    assert plot_iofv_vs_iofv(model, model)
+    iofv = model.modelfit_results.individual_ofv
+    assert plot_iofv_vs_iofv(iofv, iofv, "run1", "run1")
 
 
 def test_plot_individual_predictions(load_model_for_test, testdata):